refactor/add ng-content
