+++ src/app/features/course/course.component.ts
@@ -1,7 +1,5 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { Course } from '../courses/courses.component';
-import { faPencil, faTrash } from '@fortawesome/free-solid-svg-icons';
-import { BUTTON_ICON } from 'src/app/shared/components';
+import { Component, Input } from '@angular/core';
+import { Course } from '../courses/courses.model';
 
 @Component({
   selector: 'app-course',
@@ -9,7 +7,6 @@ import { BUTTON_ICON } from 'src/app/shared/components';
   styleUrls: ['./course.component.scss'],
 })
 export class CourseComponent {
-  @Input() canEdit = false;
   @Input() course: Course = {
     id: '',
     title: '',
@@ -18,13 +15,6 @@ export class CourseComponent {
     duration: 0,
     authors: [],
   };
-
-  @Output() showCourse = new EventEmitter();
-  @Output() editCourse = new EventEmitter();
-  @Output() deleteCourse = new EventEmitter();
-
-  public iconButtonEdit = faPencil;
-  public iconButtonDelete = faTrash;
 
   constructor() {}
 }
