+++ src/app/shared/components/header/header.component.html
@@ -1,7 +1,6 @@
 <header class="header">
   <a href="/"><img class="'logo" src="assets/svg/logo.svg" /></a>
   <div class="login">
-    <div>{{ user.name }}</div>
-    <app-button [text]="buttonText" (buttonClick)="handleLogout()"></app-button>
+    <ng-content></ng-content>
   </div>
 </header>
