+++ src/app/shared/components/info/info.component.html
@@ -1,7 +1,7 @@
 <div class="container">
-  <h2 class="title">{{ title }}</h2>
-  <p class="message" [innerHTML]="message"></p>
-  <app-button
-    (buttonClick)="buttonClick.emit()"
-    [text]="buttonText"></app-button>
+  <ng-template [ngIf]="isShowMessage">
+    <h2 class="title">{{ option.title }}</h2>
+    <p class="message" [innerHTML]="option.message"></p
+  ></ng-template>
+  <ng-content></ng-content>
 </div>
