+++ src/app/shared/components/info/info.component.ts
@@ -1,15 +1,21 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
+interface IInfo {
+  title: string;
+  message: string;
+}
+
 @Component({
   selector: 'app-info',
   templateUrl: './info.component.html',
   styleUrls: ['./info.component.scss'],
 })
 export class InfoComponent {
-  @Input() title = '';
-  @Input() message = '';
-  @Input() buttonText = '';
-  @Output() buttonClick = new EventEmitter();
+  @Input() option: IInfo = {
+    title: '',
+    message: '',
+  };
+  @Input() isShowMessage = true;
 
   constructor() {}
 }
