+++ src/app/shared/components/search/search.component.html
@@ -1,0 +1,1 @@
+<p>search works!</p>
