+++ src/app/shared/shared.module.ts
@@ -3,13 +3,17 @@ import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './components';
 import { ButtonComponent } from './components';
 import { InfoComponent } from './components/info/info.component';
-import { DataService } from './data.service';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { SearchComponent } from './components/search/search.component';
 
 @NgModule({
-  declarations: [HeaderComponent, ButtonComponent, InfoComponent],
+  declarations: [
+    HeaderComponent,
+    ButtonComponent,
+    InfoComponent,
+    SearchComponent,
+  ],
   imports: [CommonModule, FontAwesomeModule],
-  providers: [DataService],
-  exports: [HeaderComponent, ButtonComponent, InfoComponent],
+  exports: [HeaderComponent, ButtonComponent, InfoComponent, SearchComponent],
 })
 export class SharedModule {}
